Add canonical way to store firewall rules for INPUT and OUTPUT
Other chains and tables are heavily infected by docker, so they should
rather be not touched. It also means that any dockerized service that
needs firewall-based access control should use network namespace of the
host system and be running as non-root user as root user in the host
network namespace may, for example, shutdown host system.