feature(roles/adm): ssh -= {pwd-based-auth, root-login}
This diff disabeles password-based-authentication for SSH. as
well as the possibility of logging in directly as root.

While there, add to .gitignore annoying files on macOS as
well as the location where I saved the virtualevn (in many
cases I've seen virtualevens called like this, so I'd argue
this is not surprising at all).

Based on: https://ryaneschinger.com/blog/securing-a-server-with-ansible/