orchestra_nginx.conf.j2: unbreak lets encrypt
The issue at #304 is an instance of missing hooks for LE's certbot
to be able to renew because:

1. we were not listening on port 80, and

2. we were not exposing the proper directory

Do both by importing the correct j2 subtemplate.

This commit has been deployed using:

```
./play -l registry.proteus.test.ooni.io --step ./deploy-ooni-registry.yml
```

where `--step` has been used to skip reinstalling the now-gone
debian jessie repository after 54841f1221999b9a4b9a68747076b870232acc5e
allowed me to remove the repository from the system.

I did not deploy this fix to `registry.proteus.ooni.io` out of an
excess of caution. I believe more system possibly have this issue and
we should discuss with @hellais how to systematically fix it.

Note: I am sure the issue has been fixed because

```
sudo certbot renew --dry-run
```

This means that we _should_ probably create an alert that fires
if the above command fails for a host, right?

Thanks to @darkk for nudging me in the right direction.