Drop b.th.echo.ooni.io
This host was sponsored by bytemark so it runs on a very different stack
and has always been a pain to maintain since it's different to the gh
hosts.

We stopped giving out this address from the bouncer a while ago and we
confirmed that it's not hardcoded in any app hence we are dropping it.