Make the daily task run hourly
The batch pipeline reads files from the /srv/collector/renamed directly
instead of directly /srv/collector/archive becuase the daily-tasks
script does some validation and formatting of the filenames to prepare
them for the pipeline.
This task used to be running at 00:30 each day, but sometimes it did not
manage to process all the files in time for the batch pipeline (which
ticks at 01:15) resulting in files staying in the renamed dir for 24h
extra.

By doing it hourly on an ongoing basis we are able to ensure the batch
pipeline will be able to process all the collector backlog in time.

This will, however, have an impact on the fastpath which expects to read
files from /srv/collector/archive.