tor_targets.json: clearly tag directory authorities or ports
I have confirmed by reading `tor(1)` for v0.4.2.5 and by manually
invoking tor that clients use the or port and instead relays "usually"
use the dir port directly.

This means that the metric "can connect to or port" of directory
authority is different from "can connect to or port" of a specific
vanilla bridge and is different from "can connect to the dir
port" of a directory authority.

Hence, we need a way to separate these metrics. This is actually
reasonably easy because we just need to introduce a new type that
indicates "the or port of a directory authority".

This is part of https://github.com/ooni/probe/issues/973.