Derive key from IP:port || fingerprint instead.
If a Tor bridge exposes both an IPv4 and IPv6 port, we end up with the
same key because both end points share a fingerprint.  This patch
derives the key from IP:port || fingerprint instead, to prevent
collisions from happening.

For example, the key of the following bridge:

  1.2.3.4:1234 1234567890123456789012345678901234567890

will be:

  echo -n '1.2.3.4:12341234567890123456789012345678901234567890' | sha256sum
  19bc59fa00d29729edea87368ceb062b9fad66759c0fcd6b7f8680ca14452fb8